Fix for debug builds crashing (explanation below) Some reason having the UI_MenuUpdateHook inside Ui.cpp causes crashes. A while ago I noticed that it crashed when I tried moving it outside of ElDorito.cpp, but after ? moved it to Ui.cpp the crashes seemed to stop. Maybe because a lot of things were moved around when it was initially moved to Ui.cpp the crash was subsided, but now after doing the modules refactor it's returned. Only just thought to try moving it back to ElDorito.cpp and lo-and-behold it stopped the crashes. No idea why moving it outside of ElDorito.cpp would cause it to crash.
Note that I tried to clean it up after this fix and remove the redundant code (eg. making the UI_MenuUpdateHook func and the Patches::Ui::Tick func call a function which has the common code), but doing that would bring the crash back.
Crash must be caused by either the OpenUIDialogById call or the UI notification.