Removed key bindings for UI buttons because the input system needs work
Specifically, this causes a lot of issues with the chat overlay because of the huge bug where pressing Esc/Enter to close it still sends the keypress to the game. We need to think up a better way of handling input for the console. We'll likely need to use the game input functions in some way, but that's tricky in itself because we need a way of stopping the game from getting input while letting us still read it. Maybe some sort of system where we can switch in different "input contexts" might work nicely.