Improved kick commands and finished implementing IP bans
This adds and upgrades a number of server commands for kicking and banning players.
Currently, all bans are done by IP address. The ban list is checked as soon as a player tries to connect.
Additionally, player name sanitization was made stricter in order to make kicking by name easier.

What's important to note here is that there are separate commands now for kicking/banning by name, UID, and index.
This was done intentionally in order to prevent ambiguities when searching for a player.
Otherwise, for example, if you type "kick 3" and there's a player with the name "3", the wrong player might get kicked!

For kicking players, the following commands are available:

- Server.KickPlayer/kick <player name> - Kicks a player by name. Quotes are no longer necessary in order to kick players with spaces in their names.
- Server.KickUid/kickuid <uid> - Kicks a player by UID.
- Server.KickIndex/kickidx <index> - Kicks a player by index (use Server.ListPlayers/list to get this).

For banning players, the following commands are available:

- Server.KickBanPlayer/kb <player name> - Kicks a player by name and adds their IP to the ban list.
- Server.KickBanUid/kbuid <uid> - Kicks a player by UID and adds their IP to the ban list.
- Server.KickBanIndex/kbidx <index> - Kicks a player by index and adds their IP to the ban list.

Ban information is stored in a file named banlist.txt. You can use these commands to edit it:

- Server.AddBan/addban <type> <value> - Adds an entry to the ban list (but does NOT kick anyone). The only type supported for now is "ip".
- Server.Unban/unban <type> <value> - Removes an entry from the ban list. The only type supported for now is "ip".

Hopefully this will make things a lot easier for hosts!
It would be awesome if people could help me out by testing this.