Increment version number to 0.5.1.0
This does not necessarily mean that an update is coming soon. I'm just doing this because we will be making some changes to the server protocol that will make the client incompatible with existing servers. I don't want someone downloading builds from the appveyor and then complaining about how they can't join anything.