Store keyboard and controller bindings in dewrito_prefs.cfg
This overrides the game's default keyboard and controller binding system so that all bindings can be edited from dewrito_prefs.cfg. Bindings will no longer be stored in preferences.dat. This makes it possible to rebind actions (such as dual wielding) which aren't in the game menu, and it should hopefully also fix some major issues that people have been experiencing with their controls.

I haven't tested the controller options because I don't have a controller, so it would be very much appreciated if someone could do that and let me know how it works.