Replaced the controls shown in the keyboard settings menu
This lets you configure keys such as dual-wielding from inside the game without needing to edit dewrito_prefs.cfg :)

You will need to edit tags.dat and add strings for the new settings. The latest build of tagtool is required for this due to a bug that I had to fix.

Also, we will be editing the SSL scripts included with 0.5.1.0 in order to allow keys to be bound to multiple actions (needed for some of the dual wield stuff).