Proper UI keyboard input
This fixes a lot of problems with the Halo 3 UI, such as the problem where Forge items would get double-spawned or where some menus would navigate too quickly. You can even change teams in a lobby with the keyboard now by using Q and E in place of LB and RB.

It works by disabling the custom keyboard UI input code in the game and then fixing several bugs in the game's keyboard input system so that UI actions get processed correctly.