DataArray and DataIterator type safety
This renames the old "unsafe" structs to DataArrayBase and DataIteratorBase (for compatibility with engine functions which aren't type-safe) and adds new templated DataArray<>, DataIterator<>, and ConstDataIterator<> structs which hide the base methods and replace them with versions that return data of the specified type. This makes it easy to pass around data array pointers without needing to remember what type to cast data to.

Additonally, DataIterator<> and ConstDataIterator<> implement ForwardIterator semantics, and DataArray<> has begin() and end() methods, so we can use range-based for loops to iterate over data arrays! Check out src/Patches/Forge.cpp for an example.