Security, voip rejoin and other fixes
Prevents leaking voip client ips to modified dlls.
Prevents modified dlls from kicking any user.
Verifies a user connecting to teamspeak is a user in the game and is not
banned. This is based on ip.
On client kick, clears the member list so that usernames are not stuck
in the corner.
voip client shutdown will also clear the memberlist.
Shutdown of teamspeak client on status disconnected.
Added a voip.retry command to attemp to reconnect to ts while in-game so
the user will not need to rejoin.