Use a proper Cache-Control if m_ShouldCache is false
This fixes some issues with screen code getting cached when it shouldn't be, making testing difficult because reloads don't work correctly. This also removes the Reload button from the console because it's not something that end users should need to care about. Use webdebug mode if you need to work on the console.