(BREAKING) Use Promises instead of async JS callbacks
THIS WILL BREAK MOST CUSTOM SCREEN CODE. PLEASE READ THIS CAREFULLY. (Screens which are already a part of this repository have been updated.)

This changes the asynchronous JS API methods (dew.command(), etc.) to return a Promise instead of accepting onSuccess and onFailure callbacks. The main reason for this is that it allows for a lot more control over how asynchronous methods are run. It is often quite useful to run a set of asynchronous methods at the same time and then wait for all of them to finish before acting on their results (in fact, we need to be able to do this in order to fix some bugs that I encountered). There was no easy way to do this with the previous API, and using Promises makes this trivial. Additionally, this is more idiomatic with modern JS practices, and it greatly simplifies the documentation (because there's no need to document callback functions anymore).

The documentation at https://eldewrito.github.io/jsdoc/ has been updated to reflect this change. Please read through it (especially the updated "Common Tasks" tutorial) to make sure that you understand what this commit changes.

It should be fairly easy to adapt most existing code - use .then() to register your former onSuccess handler and then use .catch() to register your former onFailure handler.