Chat flood filtering (#153)
This adds a flood filtering system to text chat. Each IP address has a "spam score" associated with it. Every time a client sends a message, a spam score for the message is computed based on its contents, and then this is added to the total spam score for the client's IP. Every second, everyone's spam score is decreased by 1. If a client's total spam score reaches the maximum spam score, then the IP is timed out for a set amount of time. During the timeout period, messages sent by clients are rejected.

Currently, the only metric used to compute message spam scores is the message length. Longer messages will have a larger spam score. It would be possible to add some other metrics if necessary, e.g. increasing score based on the number of capitalized letters or increasing score based on similarity to previous messages.

All of the flood filter parameters can be controlled with console variables. Their default values are kind of arbitrary and could use some fine tuning:
- Server.FloodFilterEnabled - Controls whether or not flood filtering is enabled (default 1)
- Server.FloodMessageScoreShort - Messages which are shorter will have a score closer to this (default 2)
- Server.FloodMessageScoreLong - Messages which are longer will have a score closer to this (default 4)
- Server.FloodTimeoutScore - Once an IP reaches this score, it will be timed out (default 10)
- Server.FloodTimeoutSeconds - The number of seconds to time out an IP for (default 120)