Ignore mid-session name changes + some network code cleanup
This patches a trick that allows a player to change their name in the middle of a session by setting Player.Name and then changing teams. This check is completely server-sided. Hosts are still able to use the trick if they wish. (Perhaps setting Player.Name should immediately force a player-properties update in that case.)