Added start/stop lsnd commands, and some VoIP preference commands.
Added a command to toggle between displaying who's talking on the overlay or on the HUD. Defaults to overlay.
Replaced all instances of GetTagAddressImpl with TagInstance::IsLoaded.
Fixed a bunch of isseus with Voice Chat icons on the HUD. No known issues remain.
Added a command to toggle the Push-To-Talk sound effect.
Begun to refactor the voip related code in the Ui patchm fixed all related known bugs.
Added a couple of commands to start and stop looping sounds. Thanks to unk for the stop looping sound pointer :)