Fix map variant placement precision
Fixes an age old problem that originates with halo 3. Bungie used a
technique known as point-quanatization to compress floating point
numbers for sending accross the network. This includes object
positions/rotations among other things. The problem with this is that
you lose precision and as a result objects tend to get displaced. This
can be hard to spot, but with recent forge additions, it has become
painfully apparent.