Implement a LAN server detection backend and frontend.
This replaces the existing halo 3 LAN detection system with a new custom one using LAN broadcasting and CEF

New system
Server broadcasts a JSON encoded message containing it's infoserver port to 11773 every 1.5 seconds

Client starts a listener on 11773 when the Lan browser UI is opened, and forwards the IP and info server port of broadcasting servers to the UI which can fetch server info from the infoserver.

Old system

Client broadcasts a message to 11774 with a short header and some random data

Server responds with a broadcast to the port the client broadcasted from with server info like game mode, map, host, player count, etc, it also includes that random data that the client sent to it.

Client receives the broadcast, checks if the random data the server sent to it matches the one it send to it, if it does it parses the data and displayes it to the user.