Fix Tumblr HTTPS mismatch
As the blog is being served by Tumblr, when you visit the blog subdomain using https, the certificate is mismatched (wrong domain). So anyone clicking on this link will get redirected to an error page. Easy fix, just direct people to the insecure page, as that's what's happening after the HTTPS connection anyway.